# back/services/gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "陪玩平台 Gateway 服务 API 文档",
        "title": "陪玩平台 API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8888",
    "basePath": "/",
    "paths": {
...
        "/api/order/rate": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "对已完成的订单进行评价，包括评分和评论",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "订单"
                ],
                "summary": "评价订单",
                "parameters": [
                    {
                        "description": "评价订单请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.RateOrderRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/types.RateOrderResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/types.BaseResp"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/types.BaseResp"
                        }
                    }
                }
            }
        },
        "/api/order/start": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "开始服务订单，将订单状态从待支付改为服务中",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "订单"
                ],
                "summary": "开始服务",
                "parameters": [
                    {
                        "description": "开始服务请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.StartOrderRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/types.StartOrderResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/types.BaseResp"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/types.BaseResp"
                        }
                    }
                }
            }
        },
        "/api/user": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取当前登录用户或指定用户的详细信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "获取用户信息",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "用户ID（可选，不传则获取当前用户）",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "用户UID（可选）",
                        "name": "uid",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "手机号（可选）",
                        "name": "phone",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/types.GetUserResponse"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/types.BaseResp"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "更新当前登录用户的个人信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "更新用户信息",
                "parameters": [
                    {
                        "description": "更新用户信息请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.UpdateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/types.UpdateUserResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/types.BaseResp"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/types.BaseResp"
                        }
                    }
                }
            }
        },
...
        "/api/user/change-password": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "修改当前登录用户的密码，需要验证旧手机号和验证码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "修改密码",
                "parameters": [
                    {
                        "description": "修改密码请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.ChangePasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/types.ChangePasswordResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/types.BaseResp"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/types.BaseResp"
                        }
                    }
                }
            }
        },
...
        "/api/user/companion/apply": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "申请成为陪玩，需要设置游戏技能和价格",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "申请成为陪玩",
                "parameters": [
                    {
                        "description": "申请陪玩请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.ApplyCompanionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/types.ApplyCompanionResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/types.BaseResp"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/types.BaseResp"
                        }
                    }
                }
            }
        },
...
        "/api/user/companions/ranking/ratings": {
            "get": {
                "description": "获取陪玩按评分排名的排行榜",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "获取陪玩评分排行榜",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码（从1开始）",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页数量",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/types.GetCompanionRatingRankingResponse"
                        }
                    }
                }
            }
        },
...
        "/api/user/forgetPassword": {
            "put": {
                "description": "通过手机号和验证码重置密码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "忘记密码",
                "parameters": [
                    {
                        "description": "忘记密码请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.ForgetPasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/types.ForgetPasswordResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/types.BaseResp"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "types.ApplyCompanionRequest": {
            "type": "object",
            "properties": {
                "bio": {
                    "description": "个人简介",
                    "type": "string"
                },
                "gameSkill": {
                    "description": "游戏技能（单个游戏名称）",
                    "type": "string"
                },
                "pricePerHour": {
                    "description": "每小时价格（帅币）",
                    "type": "integer"
                }
            }
        },
        "types.ApplyCompanionResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/types.UserInfo"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "types.BaseResp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "types.ChangePasswordRequest": {
            "type": "object",
            "properties": {
                "newPassword": {
                    "type": "string"
                },
                "oldCode": {
                    "type": "string"
                },
                "oldPhone": {
                    "type": "string"
                }
            }
        },
        "types.ChangePasswordResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "types.CompanionRankingItem": {
            "type": "object",
            "properties": {
                "avatarUrl": {
                    "description": "头像URL",
                    "type": "string"
                },
                "isVerified": {
                    "description": "是否认证",
                    "type": "boolean"
                },
                "nickname": {
                    "description": "昵称",
                    "type": "string"
                },
                "rank": {
                    "description": "排名（从1开始）",
                    "type": "integer"
                },
                "rating": {
                    "description": "评分（用于评分排名）",
                    "type": "number"
                },
                "totalOrders": {
                    "description": "总接单数（用于接单数排名）",
                    "type": "integer"
                },
                "userId": {
                    "description": "用户ID",
                    "type": "integer"
                }
            }
        },
...
        "types.ForgetPasswordRequest": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
        "types.ForgetPasswordResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/types.LoginData"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "types.GetCompanionRatingRankingData": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "rankings": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.CompanionRankingItem"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "types.GetCompanionRatingRankingResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/types.GetCompanionRatingRankingData"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "types.GetUserResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/types.UserInfo"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "types.LoginData": {
            "type": "object",
            "properties": {
                "accessToken": {
                    "description": "Access Token（短期，15分钟）",
                    "type": "string"
                },
                "expiresIn": {
                    "description": "Access Token 过期时间（秒）",
                    "type": "integer"
                },
                "refreshToken": {
                    "description": "Refresh Token（长期，7天）",
                    "type": "string"
                }
            }
        },
...
        "types.OrderInfo": {
            "type": "object",
            "properties": {
                "acceptedAt": {
                    "description": "接单时间",
                    "type": "integer"
                },
                "bossId": {
                    "description": "老板ID",
                    "type": "integer"
                },
                "cancelReason": {
                    "description": "取消原因",
                    "type": "string"
                },
                "cancelledAt": {
                    "description": "取消时间",
                    "type": "integer"
                },
                "comment": {
                    "description": "评价内容",
                    "type": "string"
                },
                "companionId": {
                    "description": "陪玩ID",
                    "type": "integer"
                },
                "completedAt": {
                    "description": "完成时间",
                    "type": "integer"
                },
                "createdAt": {
                    "description": "创建时间",
                    "type": "integer"
                },
                "durationHours": {
                    "description": "时长（小时）",
                    "type": "integer"
                },
                "gameName": {
                    "description": "游戏名称",
                    "type": "string"
                },
                "id": {
                    "description": "订单ID",
                    "type": "integer"
                },
                "orderNo": {
                    "description": "订单号",
                    "type": "string"
                },
                "paidAt": {
                    "description": "支付时间",
                    "type": "integer"
                },
                "pricePerHour": {
                    "description": "每小时价格（帅币）",
                    "type": "integer"
                },
                "rating": {
                    "description": "评分",
                    "type": "number"
                },
                "startAt": {
                    "description": "开始服务时间",
                    "type": "integer"
                },
                "status": {
                    "description": "状态：1=CREATED,2=PAID,3=ACCEPTED,4=IN_SERVICE,5=COMPLETED,6=CANCELLED,7=RATED",
                    "type": "integer"
                },
                "totalAmount": {
                    "description": "订单总价（帅币）",
                    "type": "integer"
                }
            }
        },
        "types.RateOrderRequest": {
            "type": "object",
            "properties": {
                "comment": {
                    "description": "评价内容",
                    "type": "string"
                },
                "orderId": {
                    "description": "订单ID",
                    "type": "integer"
                },
                "rating": {
                    "description": "评分（0-5）",
                    "type": "number"
                }
            }
        },
        "types.RateOrderResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/types.OrderInfo"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "types.StartOrderRequest": {
            "type": "object",
            "properties": {
                "orderId": {
                    "description": "订单ID",
                    "type": "integer"
                }
            }
        },
        "types.StartOrderResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/types.OrderInfo"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "types.UpdateUserRequest": {
            "type": "object",
            "properties": {
                "avatarUrl": {
                    "description": "头像URL",
                    "type": "string"
                },
                "bio": {
                    "description": "个人简介",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "nickname": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "role": {
                    "description": "用户角色",
                    "type": "integer"
                }
            }
        },
        "types.UpdateUserResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/types.UserInfo"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "types.UserInfo": {
            "type": "object",
            "properties": {
                "avatarUrl": {
                    "description": "头像URL",
                    "type": "string"
                },
                "balance": {
                    "description": "帅币可用余额",
                    "type": "integer"
                },
                "bio": {
                    "description": "个人简介",
                    "type": "string"
                },
                "followerCount": {
                    "description": "粉丝数",
                    "type": "integer"
                },
                "followingCount": {
                    "description": "关注数",
                    "type": "integer"
                },
                "frozenBalance": {
                    "description": "冻结帅币余额（预留）",
                    "type": "integer"
                },
                "id": {
                    "description": "用户ID",
                    "type": "integer"
                },
                "nickname": {
                    "description": "用户昵称",
                    "type": "string"
                },
                "phone": {
                    "description": "手机号码",
                    "type": "string"
                },
                "role": {
                    "description": "用户角色：1=老板, 2=陪玩, 3=管理员",
                    "type": "integer"
                },
                "uid": {
                    "description": "用户唯一标识",
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT Token，格式: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}